Header and Porjects
Almost finished with my portfolio. I added functionality to contact section. I fixed the collapse menu for navbar and added functionality. I slighly fixed project navigation and descirptions. All that is left is making this accessible, cleaning up code, and fixing the responsiveness. I need the smaller screen to look just as good as larger screen and navigation to properly go to sections in both large and small screens.